fix status record

+++ finality-provider/service/fp_manager.go
@@ -135,7 +135,7 @@ func (fpm *FinalityProviderManager) monitorStatusUpdate() {
 			fpis := fpm.ListFinalityProviderInstances()
 			for _, fpi := range fpis {
 				oldStatus := fpi.GetStatus()
-				fpm.metrics.RecordFpStatus(fpi.GetBtcPkHex(), float64(oldStatus))
+				fpm.metrics.RecordFpStatus(fpi.GetBtcPkHex(), oldStatus)
 				power, err := fpi.GetVotingPowerWithRetry(latestBlock.Height)
 				if err != nil {
 					fpm.logger.Debug(
@@ -150,7 +150,7 @@ func (fpm *FinalityProviderManager) monitorStatusUpdate() {
 				if power > 0 {
 					if oldStatus != proto.FinalityProviderStatus_ACTIVE {
 						fpi.MustSetStatus(proto.FinalityProviderStatus_ACTIVE)
-						fpm.metrics.RecordFpStatus(fpi.GetBtcPkHex(), float64(proto.FinalityProviderStatus_ACTIVE))
+						fpm.metrics.RecordFpStatus(fpi.GetBtcPkHex(), proto.FinalityProviderStatus_ACTIVE)
 						fpm.logger.Debug(
 							"the finality-provider status is changed to ACTIVE",
 							zap.String("fp_btc_pk", fpi.GetBtcPkHex()),
@@ -172,7 +172,7 @@ func (fpm *FinalityProviderManager) monitorStatusUpdate() {
 				// power == 0 and slashed == true, set status to SLASHED and stop and remove the finality-provider instance
 				if slashed {
 					fpm.setFinalityProviderSlashed(fpi)
-					fpm.metrics.RecordFpStatus(fpi.GetBtcPkHex(), float64(proto.FinalityProviderStatus_SLASHED))
+					fpm.metrics.RecordFpStatus(fpi.GetBtcPkHex(), proto.FinalityProviderStatus_SLASHED)
 					fpm.logger.Debug(
 						"the finality-provider is slashed",
 						zap.String("fp_btc_pk", fpi.GetBtcPkHex()),
@@ -183,7 +183,7 @@ func (fpm *FinalityProviderManager) monitorStatusUpdate() {
 				// power == 0 and slashed_height == 0, change to INACTIVE if the current status is ACTIVE
 				if oldStatus == proto.FinalityProviderStatus_ACTIVE {
 					fpi.MustSetStatus(proto.FinalityProviderStatus_INACTIVE)
-					fpm.metrics.RecordFpStatus(fpi.GetBtcPkHex(), float64(proto.FinalityProviderStatus_INACTIVE))
+					fpm.metrics.RecordFpStatus(fpi.GetBtcPkHex(), proto.FinalityProviderStatus_INACTIVE)
 					fpm.logger.Debug(
 						"the finality-provider status is changed to INACTIVE",
 						zap.String("fp_btc_pk", fpi.GetBtcPkHex()),
@@ -244,7 +244,7 @@ func (fpm *FinalityProviderManager) StartAll() error {
 	}
 
 	for _, fp := range storedFps {
-		fpm.metrics.RecordFpStatus(fp.GetBIP340BTCPK().MarshalHex(), float64(fp.Status))
+		fpm.metrics.RecordFpStatus(fp.GetBIP340BTCPK().MarshalHex(), fp.Status)
 		if fp.Status == proto.FinalityProviderStatus_CREATED || fp.Status == proto.FinalityProviderStatus_SLASHED {
 			fpm.logger.Info("the finality provider cannot be started with status",
 				zap.String("btc-pk", fp.GetBIP340BTCPK().MarshalHex()),
